refactor: move throttler type into its own subPackage
