+++ rpc/v8/simulation.go
@@ -14,7 +14,7 @@ import (
 	"github.com/NethermindEth/juno/core/felt"
 	"github.com/NethermindEth/juno/jsonrpc"
 	"github.com/NethermindEth/juno/rpc/rpccore"
-	"github.com/NethermindEth/juno/utils"
+	"github.com/NethermindEth/juno/utils/throttler"
 	"github.com/NethermindEth/juno/vm"
 	"github.com/consensys/gnark-crypto/ecc/stark-curve/fp"
 )
@@ -197,7 +197,7 @@ func (h *Handler) prepareTransactions(
 			network,
 		)
 		if aErr != nil {
-			if errors.Is(aErr, utils.ErrResourceBusy) {
+			if errors.Is(aErr, throttler.ErrResourceBusy) {
 				return nil, nil, rpccore.ErrInternal.CloneWithData(rpccore.ThrottledCompilerErr)
 			}
 			return nil, nil, jsonrpc.Err(jsonrpc.InvalidParams, aErr.Error())
@@ -213,7 +213,7 @@ func (h *Handler) prepareTransactions(
 }
 
 func handleExecutionError(err error) *jsonrpc.Error {
-	if errors.Is(err, utils.ErrResourceBusy) {
+	if errors.Is(err, throttler.ErrResourceBusy) {
 		return rpccore.ErrInternal.CloneWithData(rpccore.ThrottledVMErr)
 	}
 	var txnExecutionError vm.TransactionExecutionError
