+++ utils/throttler/throttler.go
@@ -1,4 +1,4 @@
-package utils
+package throttler
 
 import (
 	"errors"
